support floating point scale (#27)

+++ Source/gltf2usd.py
@@ -1119,7 +1119,7 @@ if __name__ == '__main__':
     parser.add_argument('--fps', action='store', dest='fps', help='The frames per second for the animations', type=float, default=24)
     parser.add_argument('--output', '-o', action='store', dest='usd_file', help='destination to store generated .usda file', required=True)
     parser.add_argument('--verbose', '-v', action='store_true', dest='verbose', help='Enable verbose mode')
-    parser.add_argument('--scale', '-s', action='store', dest='scale', help='Scale the resulting USDA', type=int, default=100)
+    parser.add_argument('--scale', '-s', action='store', dest='scale', help='Scale the resulting USDA', type=float, default=100)
     args = parser.parse_args()
 
     if args.gltf_file:
